saperated FormComponent
