+++ client/src/components/App/Content/InvestigationForm/TabManagement/InteractionsTab/InteractionDialog/InteractionEventForm/ContactsSection/ContactsTabs.tsx
@@ -6,26 +6,8 @@ import { Divider, Tab, Tabs, useTheme } from '@material-ui/core';
 
 import useFormStyles from 'styles/formStyles';
 
+import Forms from './Forms';
 import useStyles from './ContactsTabsStyles';
-import FamilyMembersForm from './FamilyMembers/FamilyMembersForm';
-import ManualContactsForm from './ManualContactsForm/ManualContactsForm';
-
-const contactFormTabs = [
-    { id: 0, Component: <ManualContactsForm /> },
-    { id: 1, Component: <FamilyMembersForm /> },
-];
-
-const FormComponent = ({ currentTab }: { currentTab: number }) => {
-    const formClasses = useFormStyles();
-
-    return <>
-        {contactFormTabs.map(tab =>
-            <div className={currentTab !== tab.id ? formClasses.hidden : formClasses.formSize}>
-                {tab.Component}
-            </div>
-        )}
-    </>
-};
 
 const ContactsTabs = ({ isVisible }: { isVisible: boolean }) => {
     const [currentTab, setTab] = React.useState<number>(0);
@@ -57,7 +39,7 @@ const ContactsTabs = ({ isVisible }: { isVisible: boolean }) => {
                     label='בני משפחה' />
             </Tabs>
             <Divider orientation='vertical' variant='fullWidth' light={true} />
-            <FormComponent currentTab={currentTab} />
+            <Forms currentTab={currentTab} />
         </div>
     );
 };
